Remove unused DEBUG_DRAWINGS code
This code appears to be from earlier experiments reading drawing
groups (charts and images). It hasn't been touched in 4 years, and
likely belongs in a separate code branch until it's production-ready.

In the interest of simplicity I'm just going to remove the code
until there is actual demand for reading in drawings.