New --enable-fuzz-testing configure option
The newest version of clang lets you switch on the fuzz-testing run-time
with -fsanitize=fuzzer. Some of the infrastructure was already in place,
but now it's easily accessible.

MacOS users should note that the clang that ships with Xcode includes
the -fsanitize=fuzzer compiler option, but the Xcode installation does
not contain the appropriate static library. So you'll need to download
clang 7 from clang.llvm.org and use that instead.