Make C++ test conditional on having a C++11 compiler
Ancient but supported platforms like RHEL 6 ship a version of GCC that
doesn't support C++11 causing ./configure to fail on that check. But
since C++ is only used for cpp_test we can make it conditional on having
a proper compiler and skip it otherwise.

Requires autoconf-archive, now preinstalled on the Mac and Windows CI
pipeline.