Fix long-standing "extra column" bug
Turns out the ROW record uses a 0-index to reference the first column in
a row, and a 1-index to reference the last column in a row. This means
that `lastcol' has been assigned an incorrect value (N+1) since time
immemorial. Fix it, and eliminate the extra blank column in everyone's
CSV files.

Fixes #73