[ABI break] Re-use iconv_t objects for string conversions
Previously, every string conversion resulted in its own call to
iconv_open and iconv_close. This seems pretty inefficient, so instead
we cache the iconv_t objects in the workbook objects, and close them
when the workbook is closed.

For BIFF8 documents, up to two converters are needed: one for UTF-16LE
conversions, and a second converter for "compressed" UTF-16 (in which
all the high bytes are 0), which I understand to be the same as Latin-1.
iconv is skipped for the latter strings in favor of an in-line
implementation if the output encoding is set to UTF-8.

For BIFF5 documents, a single converter from the code page specified by
the document to the client's desired encoding is needed.

Note to self: I am still not happy with the use of wcstombs as it sets
the CTYPE locale. Need to eliminate this side effect at some point,
particularly since this change always hits wcstombs during the OLE
reading.