Fix libtool versioning
The version-info shipped in 1.6.0 was 7:0:6, whereas 1.5.3 was 6:3:5.
The AGE field should have been set to 0 to reflect ABI breakage.

I've changed the logic to set AGE to 0 always, since I very much doubt
public interfaces in this library will be added without other breakage.
And also because it is annoying to keep up with libtool's versioning
rules.

In the future we should remove some of the private functions from the
public API but I think we've had enough damage for one week.

See #80